Create and include event-host-list block
